requirements.txt
aiohttp>=3.7.4,<4
discord.py>=2.3.1
python-dotenv
deep_translator
